Refactor file system and statememory
