+++ js/api.ts
@@ -3,6 +3,7 @@ import { ModelFormat } from "./io/format";
 import { Prop } from "./misc";
 import { EventSystem } from "./util/event_system";
 import { compareVersions } from "./util/util";
+import { FileSystem } from "./file_system";
 
 declare const appVersion: string;
 declare let Format: ModelFormat
@@ -342,7 +343,19 @@ export const Blockbench = {
 	Project: 0 as (ModelProject | number),
 	get Undo() {
 		return Project?.undo;
-	}
+	},
+	// File System
+	import: FileSystem.importFile,
+	importFile: FileSystem.importFile,
+	pickDirectory: FileSystem.pickDirectory,
+	read: FileSystem.readFile,
+	readFile: FileSystem.readFile,
+	export: FileSystem.exportFile,
+	exportFile: FileSystem.exportFile,
+	writeFile: FileSystem.writeFile,
+	findFileFromContent: FileSystem.findFileFromContent,
+	addDragHandler: FileSystem.addDragHandler,
+	removeDragHandler: FileSystem.removeDragHandler,
 };
 
 (function() {
@@ -369,53 +382,8 @@ if (isApp) {
 	if (Blockbench.platform.includes('win32') === true) window.osfs = '\\';
 }
 
-
-
-export const StateMemory = {
-	/**
-	 * Initialize a memorized property
-	 */
-	init(key: string, type: 'string'|'number'|'boolean'|'object'|'array') {
-		let saved: any = localStorage.getItem(`StateMemory.${key}`)
-		if (typeof saved == 'string') {
-			try {
-				saved = JSON.parse(saved)
-			} catch (err) {
-				localStorage.removeItem(`StateMemory.${key}`);
-			}
-		}
-		if ( saved !== null && (typeof saved == type || (type == 'array' && (saved instanceof Array))) ) {
-			StateMemory[key] = saved;
-		} else {
-			StateMemory[key] = (() => {switch (type) {
-				case 'string': return ''; break;
-				case 'number': return 0; break;
-				case 'boolean': return false; break;
-				case 'object': return {}; break;
-				case 'array': return []; break;
-			}})();
-		}
-	},
-	set(key: string, value) {
-		if (StateMemory[key] instanceof Array) {
-			StateMemory[key].replace(value);
-		} else {
-			StateMemory[key] = value;
-		}
-		StateMemory.save(key);
-	},
-	save(key: string) {
-		let serialized = JSON.stringify(StateMemory[key])
-		localStorage.setItem(`StateMemory.${key}`, serialized)
-	},
-	get(key: string): string|number|[]|boolean|any {
-		return StateMemory[key];
-	}
-}
-
 Object.assign(window, {
 	LastVersion,
 	Blockbench,
-	StateMemory,
 	isApp
 });
